role permission set

+++ app/Http/Controllers/RoleController.php
@@ -16,38 +16,52 @@ class RoleController extends BaseController
 
     public function index()
     {
-        $this->setPageData('Role','Role','fas fa-th-list');
-        return view('role.index');
+        if (permission('role-access')) {
+            $this->setPageData('Role','Role','fas fa-th-list');
+            return view('role.index');
+        } else {
+            return $this->unauthorized_access_blocked();
+        }
     }
 
     public function getDatatableData(Request $request)
     {
-        if($request->ajax()){
-            $output = $this->service->getDatatableData($request);
-        }else{
-            $output = ['status' => 'error', 'message' => 'Unauthorized action blocked!'];
-        }
+        if (permission('role-access')) {
+            if($request->ajax()){
+                $output = $this->service->getDatatableData($request);
+            }else{
+                $output = ['status' => 'error', 'message' => 'Unauthorized action blocked!'];
+            }
 
-        return response()->json($output);
+            return response()->json($output);
+        }
     }
 
     public function create()
     {
-        $this->setPageData('Create Role','Create Role','fas fa-th-list');
-        $data = $this->service->PermissionModuleList();
-        return view('role.create', compact('data'));
+        if (permission('role-add')) {
+            $this->setPageData('Create Role','Create Role','fas fa-th-list');
+            $data = $this->service->PermissionModuleList();
+            return view('role.create', compact('data'));
+        }else{
+            return $this->unauthorized_access_blocked();
+        }
     }
 
   
     public function storeOrUpdate(RoleRequest $request)
     {
         if($request->ajax()){
-            $result = $this->service->storeOrUpdateData($request);
-            if($result){
-                return $this->response_json($status='success',$message='Data Has Been Saved Successfully',$data=null,$response_code=200);
+            if (permission('role-add') || permission('role-edit')) {
+                $result = $this->service->storeOrUpdateData($request);
+                if($result){
+                    return $this->response_json($status='success',$message='Data Has Been Saved Successfully',$data=null,$response_code=200);
+                }else{
+                    return $this->response_json($status='error',$message='Data Cannot Save',$data=null,$response_code=204);
+                }
             }else{
-                return $this->response_json($status='error',$message='Data Cannot Save',$data=null,$response_code=204);
-            }
+                    return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
+                }
         }else{
            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
         }
@@ -55,23 +69,32 @@ class RoleController extends BaseController
 
     public function edit(int $id)
     {
-        $this->setPageData('Edit Role','Edit Role','fas fa-th-list');
-        $data = $this->service->PermissionModuleList();
-        $permission_data = $this->service->edit($id);
-        return view('role.edit',compact('data','permission_data'));
+        if (permission('role-edit')) {
+            $this->setPageData('Edit Role','Edit Role','fas fa-th-list');
+            $data = $this->service->PermissionModuleList();
+            $permission_data = $this->service->edit($id);
+            return view('role.edit',compact('data','permission_data'));
+        }else{
+            return $this->unauthorized_access_blocked();
+        }
     }
 
     public function show(int $id)
     {
-        $this->setPageData('Role Details','Role Details','fas fa-th-list');
-        $data = $this->service->PermissionModuleList();
-        $permission_data = $this->service->edit($id);
-        return view('role.view',compact('data','permission_data'));
+        if (permission('role-view')) {
+            $this->setPageData('Role Details','Role Details','fas fa-th-list');
+            $data = $this->service->PermissionModuleList();
+            $permission_data = $this->service->edit($id);
+            return view('role.view',compact('data','permission_data'));
+        }else{
+            return $this->unauthorized_access_blocked();
+        }
     }
 
     public function delete(Request $request)
     {
         if($request->ajax()){
+            if (permission('role-delete')) {
             $result = $this->service->delete($request);
             if($result == 1){
                 return $this->response_json($status='success',$message='Data Has Been Deleted Successfully',$data=null,$response_code=200);
@@ -81,6 +104,9 @@ class RoleController extends BaseController
                 return $this->response_json($status='error',$message='Data Cannot Delete',$data=null,$response_code=204);
             }
         }else{
+            return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
+        }
+        }else{
            return $this->response_json($status='error',$message=null,$data=null,$response_code=401);
         }
     }
@@ -88,13 +114,17 @@ class RoleController extends BaseController
     public function bulkDelete(Request $request)
     {
         if($request->ajax()){
-            $result = $this->service->bulkDelete($request);
-            if($result['status'] == 1){
-                return $this->response_json($status='success',$message= !empty($result['message']) ? $result['message'] : 'Data Has Been Deleted Successfully',$data=null,$response_code=200);
-            }elseif($result['status'] == 2){
-                return $this->response_json($status='error',$message= !empty($result['message']) ? $result['message'] : 'Selected Data Cannot Delete',$data=null,$response_code=204);
+            if (permission('role-bulk-delete')) {
+                $result = $this->service->bulkDelete($request);
+                if($result['status'] == 1){
+                    return $this->response_json($status='success',$message= !empty($result['message']) ? $result['message'] : 'Data Has Been Deleted Successfully',$data=null,$response_code=200);
+                }elseif($result['status'] == 2){
+                    return $this->response_json($status='error',$message= !empty($result['message']) ? $result['message'] : 'Selected Data Cannot Delete',$data=null,$response_code=204);
+                }else{
+                    return $this->response_json($status='error',$message= !empty($result['message']) ? $result['message'] : 'Selected Data Cannot Delete',$data=null,$response_code=204);
+                }
             }else{
-                return $this->response_json($status='error',$message= !empty($result['message']) ? $result['message'] : 'Selected Data Cannot Delete',$data=null,$response_code=204);
+                return $this->response_json($status='error',$message='Unauthorized Access Blocked',$data=null,$response_code=401);
             }
         }else{
             return $this->response_json('error',null,null,401);
